requirements.txt
langchain
openai
chromadb
fastapi
uvicorn
pypdf
python-dotenv
streamlit
pillow
pytesseract
opencv-python
pyzbar
